% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_egm_line.R
\name{find_egm_line}
\alias{find_egm_line}
\title{Read single line of EGM5 data.}
\usage{
find_egm_line(line)
}
\arguments{
\item{line}{Vector (or row) of EGM data.}
}
\value{
A data.frame that includes one row of EGM5 data with correct column names.
}
\description{
This function reads one line of EGM
}
\examples{
line_example <- "M1,2024-03-05,12:00,Plot1,1,400,1013,200,10,25,21,0"
line_processed <- find_egm_line(line_example)
}
\keyword{co2}
\keyword{egm}
\keyword{egm5}
\keyword{setting}
