% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_egm_setting.R
\name{find_egm_setting}
\alias{find_egm_setting}
\title{Find setting of EGM5}
\usage{
find_egm_setting(filename)
}
\arguments{
\item{filename}{Character that contains the data.}
}
\value{
A character representing the measurement setting of EGM in a specific line.
}
\description{
This function reads out the setting of EGM5 for a specific line.
}
\examples{
# Example usage of the function
setting_line_1 <- find_egm_setting(filename)
}
\keyword{co2}
\keyword{egm}
\keyword{egm5}
\keyword{setting}
