% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/night_highlights.R
\name{highlight_night}
\alias{highlight_night}
\title{Highlight Night Function}
\usage{
highlight_night(
  data = "data",
  lat = "lat",
  datetime_name = "datetime",
  hcolor = "blue",
  halpha = 0.2
)
}
\arguments{
\item{data}{Data frame containing a datetime column.}

\item{lat}{Latitude for calculating day/night.}

\item{datetime_name}{Name of the datetime column, defaults to "datetime"}

\item{hcolor}{Color for highlighting, defaults to "blue"}

\item{halpha}{Alpha transparency for highlighting, defaults to 0.2}
}
\description{
This function highlights night periods in a ggplot.
}
\note{
Requires the ggplot2, LakeMetabolizer, lubridate and dplyr packages.
}
\examples{
library(ggplot2) # Ensure ggplot2 is available
# Define the start and end date
start_date <- as.POSIXct("2024-03-01 00:00:00")
end_date <- as.POSIXct("2024-03-04 23:00:00")

# Create a sequence of hourly timestamps
timestamp <- seq(from = start_date, to = end_date, by = "hour")

# Generate corresponding temperature values ranging between 10 and 25
temperature <- seq(from = 10, to = 25, length.out = length(timestamp))

# Create a data frame with timestamp and temperature columns
temperature_data <- data.frame(timestamp = timestamp, temperature = temperature)

# Plot the data with night highlights
ggplot(temperature_data) +
highlight_night(data = temperature_data, 52, datetime_name = "timestamp") +
geom_line(aes(timestamp, temperature))
}
