% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_co2t.R
\name{reaad_co2t}
\alias{reaad_co2t}
\title{Brief description of the function}
\usage{
reaad_co2t(arg1, arg2)
}
\arguments{
\item{arg1}{Description of the first argument}

\item{arg2}{Description of the second argument}
}
\value{
Description of the return value
}
\description{
More detailed description of the function.
This can span multiple lines if needed.
}
\examples{
# Example usage of the function
my_function(arg1 = "value1", arg2 = 10)

}
\keyword{keyword1}
\keyword{keyword2}
