% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_egm.R
\name{read_egm}
\alias{read_egm}
\title{Function to read data from EGM CO2 Analyzer files}
\usage{
read_egm(filename, egm_model, year)
}
\arguments{
\item{filename}{Filename of the datafile.}

\item{egm_model}{Character, which contains the model number of the EGM CO2 analyzer.}

\item{year}{Character, which contains the year of the measurements, because EGM4 does not record year. Parameter will be discarded of EGM5 was used.}
}
\value{
Clean datafile of the measurements.
}
\description{
This function can be used to read data from EGM4 and EGM5 CO2 Analyzers.
}
\keyword{co2}
\keyword{egm}
