% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gc.R
\name{read_gc}
\alias{read_gc}
\title{Function to read data from GC2 files}
\usage{
read_gc(filename)
}
\arguments{
\item{filename}{Filename of the datafile.}
}
\value{
Clean datafile of the measurements.
}
\description{
This function can be used to read data GC template.
}
\examples{
# Example usage of the function
read_gc(filename = "path\\\\to\\\\file.txt")

}
\keyword{co2}
\keyword{gc}
