% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_egm_line.R
\name{find_egm_line}
\alias{find_egm_line}
\title{Read single line of EGM5 data.}
\usage{
find_egm_line(line)
}
\arguments{
\item{filename}{Character that contains the data.}
}
\value{
A data.frame that includes one row of EGM5 data with correct column names.
}
\description{
This function reads one line of EGM
}
\examples{
# Example usage of the function
find_egm_line <- find_egm_setting(line)
}
\keyword{co2}
\keyword{egm}
\keyword{egm5}
\keyword{setting}
