% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_egm_setting.R
\name{find_egm_setting}
\alias{find_egm_setting}
\title{Function to read measurement setting of EGM 5}
\usage{
find_egm_setting(filename)
}
\arguments{
\item{filename}{Filename of the datafile.}
}
\value{
"M1" - "M6"
}
\description{
EGM 5 datafiles are different between 6 types of measurement setting. This function reads the setting from the datafile.
}
\examples{
# Example usage of the function
read_egm(filename = "path\\\\to\\\\file.txt", egm_model = "4")

}
\keyword{egm}
\keyword{setting}
