% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_egm_setting.R
\name{find_egm_setting}
\alias{find_egm_setting}
\title{Find setting of EGM5}
\usage{
find_egm_setting(filename)
}
\arguments{
\item{filename}{Character that contains the data.}
}
\value{
A character representing the measurement setting of EGM in a specific line.
}
\description{
This function reads out the setting of EGM5 for a specific line.
}
\examples{
# Create a temporary file with example EGM5 data
temp_file <- tempfile()
writeLines(c("Zero", "M1,2024-03-05,12:00,Plot1,1,400,1013,200,10,25,21,0"), temp_file)

# Use the temporary file as input
setting_line_1 <- find_egm_setting(temp_file)

# Print the setting
print(setting_line_1)

# Clean up
unlink(temp_file)
}
\keyword{co2}
\keyword{egm}
\keyword{egm5}
\keyword{setting}
