% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_contros.R
\name{read_contros}
\alias{read_contros}
\title{Read data of Contros probes}
\usage{
read_contros(filename, tz)
}
\arguments{
\item{filename}{Character, the SD_datafile that contains the data}

\item{tz}{Character, the timezone setting which was used for the Contros probe}
}
\value{
A dataframe of the contros data
}
\description{
This function can be used to read Contros probe data from SD data files
This can span multiple lines if needed.
}
\keyword{CO2}
\keyword{contros}
