% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_egm.R
\name{read_egm}
\alias{read_egm}
\title{Function to read data from EGM CO2 Analyzer files}
\usage{
read_egm(filename, egm_model, year)
}
\arguments{
\item{filename}{Filename of the datafile.}

\item{egm_model}{Character, which contains the model number of the EGM CO2 analyzer.}
}
\value{
Clean datafile of the measurements.
}
\description{
This function can be used to read data from EGM4 and EGM5 CO2 Analyzers.
}
\examples{
# Example usage of the function
read_egm(filename = "path\\\\to\\\\file.txt", egm_model = "4")

}
\keyword{co2}
\keyword{egm}
