% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hobo.R
\name{read_hobo}
\alias{read_hobo}
\title{Read data from Hobo logger.}
\usage{
read_hobo(filename, logger_id)
}
\arguments{
\item{filename}{Filename of the datafile.}

\item{logger_id}{Character, Logger ID, or custom name.}
}
\value{
Clean datafile of the measurements.
}
\description{
This function reads data files of hobo loggers.
}
\examples{
# Example usage of the function
read_egm(filename = "path\\\\to\\\\file.txt", logger_id = "climate_chamber_1")
}
\keyword{hobo}
\keyword{temperature}
